+++ src/simple_animl/models/technique.py
